rewrites: locally escape scope
Rather than having to pull a closure out of a scope and insert it into the next
scope, possibly far away, create a new scope for it locally, which can be
permuted out and eventually merged with the next scope using other rules.

This is a partial rewrite of 02193f5: Rather than using permutations to simplify
compounding, we do the converse and use compounding to simplify permutations.
This way we get locality for all our rules (in a certain sense at least).