terms: relax restriction on distributors
We no longer require the variables listed in the medial to exist in the term it
applies to (side-conditions vii and viii). The semantics changes from specifying
which hole each wire goes through, into what wires might potentially go through
each hole.

The immediate consequence of this is that we can simplify the rewrite rules, the
eventual aim is to get the rewrite rules and the definition to a point where we
can trivially change the way we view our terms; from medials being explict
constructors into being implicitly specified by labelling wires.